% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{agi}
\alias{agi}
\title{Calculate AGI variables from individual-year-of-life data}
\usage{
agi(my_iyol, members_l, focals_l, females_l, agonism_l, min_res_days = 60,
  parallel = FALSE, ncores = NULL)
}
\arguments{
\item{my_iyol}{Individual-year-of-life data.}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focals_l}{A subset of focals produced by the function 'subset_focals'}

\item{females_l}{A subset of female counts produced by the function 'subset_females'}

\item{agonism_l}{A subset of agonism data produced by the function 'subset_agonism'}

\item{min_res_days}{The minimum number of coresidence days needed for dyad to be included. Defaults to 60 days.}
}
\value{
The input data with an additional list columsn containing the full AGI subset and variables.
}
\description{
Calculate AGI variables from individual-year-of-life data
}
