% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\docType{data}
\name{amboseli_palettes}
\alias{amboseli_palettes}
\title{List of Amboseli color palettes}
\format{An object of class \code{list} of length 4.}
\usage{
amboseli_palettes
}
\description{
Use \code{\link{amboseli_palette}} to select palettes of desired length.
}
\keyword{datasets}
