% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{dsi_summary}
\alias{dsi_summary}
\title{Create summary data from DSI input}
\usage{
dsi_summary(df)
}
\arguments{
\item{df}{DSI input data produced by the 'dsi' function.}
}
\value{
An individual-year-of-life data set with summarized DSI variables.
}
\description{
Create summary data from DSI input
}
