% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{dyadic_index}
\alias{dyadic_index}
\title{Calculate dyadic index variables from individual-year-of-life data}
\usage{
dyadic_index(my_iyol, biograph_l, members_l, focals_l, females_l,
  interactions_l, min_cores_days = 60, within_grp = FALSE,
  parallel = FALSE, ncores = NULL, directional = FALSE)
}
\arguments{
\item{my_iyol}{Individual-year-of-life data.}

\item{biograph_l}{A local copy of the biograph table}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focals_l}{A subset of focals produced by the function 'subset_focals'}

\item{females_l}{A subset of female counts produced by the function 'subset_females'}

\item{min_cores_days}{The minimum number of coresidence days needed for dyad to be included. Defaults to 60 days.}

\item{within_grp}{Logical value indicating whether regressions should be fit relative to dyads in entire population (default) or to dyads within-group}

\item{parallel}{Logical value indicating whether to process in parallel}

\item{ncores}{Integer value indicating how many cores to use in parallel processing}

\item{directional}{Logical value indicating whether to preserve directionality}

\item{agonism_l}{A subset of agonism data produced by the function 'subset_agonism'}
}
\value{
The input data with an additional list columsn containing the full DSI subset and the focal DSI variables.
}
\description{
Calculate dyadic index variables from individual-year-of-life data
}
