% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{dyadic_index_summary}
\alias{dyadic_index_summary}
\title{Create summary data from dyadic index input}
\usage{
dyadic_index_summary(df)
}
\arguments{
\item{df}{Dyadic index input data produced by the 'dyadic_index' function.}
}
\value{
An individual-year-of-life data set with summarized dyadic variables.
}
\description{
Create summary data from dyadic index input
}
