% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{dyadic_row_summary}
\alias{dyadic_row_summary}
\title{Summarize a single DSI subset.
Calculate top partners, bond quantity, bond strength, and bond reciprocity
for each individual-year-of-life separately for the focal animal's top 3
grooming partners, separately for F-M and F-F dyads (if applicable)}
\usage{
dyadic_row_summary(df, focal, directional)
}
\arguments{
\item{df}{A DSI subset}

\item{focal}{Focal animal}

\item{directional}{Logical value indicating whether to preserve directionality}
}
\value{
The input data with additional list columns.
}
\description{
Summarize a single DSI subset.
Calculate top partners, bond quantity, bond strength, and bond reciprocity
for each individual-year-of-life separately for the focal animal's top 3
grooming partners, separately for F-M and F-F dyads (if applicable)
}
