% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{fit_dsi_regression}
\alias{fit_dsi_regression}
\title{Fit DSI regression on subset of data}
\usage{
fit_dsi_regression(df)
}
\arguments{
\item{df}{A subset of data on which to fit a regression of grooming on observer effort.}
}
\value{
The input data with an additional column containing the regression residuals.
}
\description{
Fit DSI regression on subset of data
}
