% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{fit_dyadic_regression}
\alias{fit_dyadic_regression}
\title{Fit dyadic index regression on subset of data}
\usage{
fit_dyadic_regression(df)
}
\arguments{
\item{df}{A subset of data on which to fit a regression of interactions on observer effort.}
}
\value{
The input data with an additional column containing the regression residuals.
}
\description{
Fit dyadic index regression on subset of data
}
