% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{get_agi_subset}
\alias{get_agi_subset}
\title{Obtain composite agonism subsets for the animal's year of life}
\usage{
get_agi_subset(df, members_l, focals_l, females_l, agonism_l,
  min_res_days = 60)
}
\arguments{
\item{df}{One row individual-year-of-life data}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focals_l}{A subset of focals produced by the function 'subset_focals'}

\item{females_l}{A subset of female counts produced by the function 'subset_females'}

\item{agonism_l}{A subset of agonism data produced by the function 'subset_agonism'}

\item{min_res_days}{The minimum number of residence days needed to be included. Defaults to 60 days.}
}
\value{
The input row with an additional list column containing the subset
}
\description{
Obtain composite agonism subsets for the animal's year of life
}
