% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{get_focal_dsi}
\alias{get_focal_dsi}
\title{Calculate DSI variables for focal animal from a DSI subset}
\usage{
get_focal_dsi(my_sname, my_grp, my_subset)
}
\arguments{
\item{my_subset}{}
}
\value{
The input data with an additional list column containing the DSI variables.
}
\description{
Calculate DSI variables for focal animal from a DSI subset
}
