% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{get_focal_index}
\alias{get_focal_index}
\title{Calculate dyadic variables for focal animal from a dyadic subset}
\usage{
get_focal_index(my_sname, my_grp, my_subset)
}
\arguments{
\item{my_subset}{}
}
\value{
The input data with an additional list column containing the dyadic variables.
}
\description{
Calculate dyadic variables for focal animal from a dyadic subset
}
