% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-membership.R
\name{get_n_members}
\alias{get_n_members}
\title{get_n_members}
\usage{
get_n_members(biograph, members, my_grp, my_date, my_sex = NULL)
}
\arguments{
\item{biograph}{A DBI connection to the babase table biograph, or an equivalent biograph table in your R environment}

\item{members}{A DBI connection to the babase table members, or an equivalent biograph table in your R environment}

\item{my_grp}{The group of interest}

\item{my_date}{The date of interest}

\item{my_sex}{The sex of interest. If null, assumes all individuals.}
}
\value{
A count of individuals in the group on that date.
}
\description{
get_n_members
}
