% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{make_iyol}
\alias{make_iyol}
\title{Create a data frame for each adult individual year of life}
\usage{
make_iyol(babase, members_l, focalcount, grooming_interactions)
}
\arguments{
\item{babase}{A DBI connection to the babase database}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focalcount}{A subset of focals produced by the function 'subset_focals'}

\item{grooming_interactions}{A subset of grooming data produced by the function 'subset_grooming'}
}
\value{
A tibble with one row per animal and year of life, with contextual data
}
\description{
Create a data frame for each adult individual year of life
}
