% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{make_palette}
\alias{make_palette}
\title{A color palette generator}
\usage{
make_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{amboseli_div}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colors.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
A color palette generator
}
\examples{
make_palette("amboseli_div")
make_palette("amboseli_div", 5)

# If you need more colors than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colors
pal <- make_palette(name = "amboseli_div", n = 100, type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
