% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{sci}
\alias{sci}
\title{Calculate SCI variables from individual-year-of-life data}
\usage{
sci(my_iyol, members_l, focalcount, f_count, grooming_interactions,
  min_res_days = 60)
}
\arguments{
\item{my_iyol}{Individual-year-of-life data.}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focalcount}{A subset of focals produced by the function 'subset_focals'}

\item{f_count}{A subset of female counts produced by the function 'subset_females'}

\item{grooming_interactions}{A subset of grooming data produced by the function 'subset_grooming'}

\item{min_res_days}{The minimum number of coresidence days needed for dyad to be included. Defaults to 60 days.}
}
\value{
The input data with an additional list columsn containing the full SCI subset and variables.
}
\description{
Calculate SCI variables from individual-year-of-life data
}
