% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{subset_agonism}
\alias{subset_agonism}
\title{Obtain a subset of agonism data that excludes behavioral observation gaps.}
\usage{
subset_agonism(babase, members_l)
}
\arguments{
\item{babase}{A DBI connection to the babase database}

\item{members_l}{A subset of members table produced by the function 'subset_members'}
}
\value{
A subset of agonsism data that excludes behavioral observation gaps.
}
\description{
Obtain a subset of agonism data that excludes behavioral observation gaps.
}
