% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biographical-data.R
\name{subset_females}
\alias{subset_females}
\title{Obtain a subset of adult female count data that excludes flagged behavioral observation periods.}
\usage{
subset_females(members_l)
}
\arguments{
\item{members_l}{A subset of members table produced by the function 'subset_members'}
}
\value{
A subset of female count data that excludes flagged behavioral observation periods.
}
\description{
Obtain a subset of adult female count data that excludes flagged behavioral observation periods.
}
