% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biographical-data.R
\name{subset_focals}
\alias{subset_focals}
\title{Obtain a subset of adult female focal samples that excludes flagged behavioral observation periods.}
\usage{
subset_focals(babase, members_l)
}
\arguments{
\item{babase}{A DBI connection to the babase database}

\item{members_l}{A subset of members table produced by the function 'subset_members'}
}
\value{
A subset of focal samples that excludes flagged behavioral observation periods.
}
\description{
Obtain a subset of adult female focal samples that excludes flagged behavioral observation periods.
}
