% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biographical-data.R
\name{subset_members}
\alias{subset_members}
\title{Obtain a subset of members table that excludes flagged behavioral observation periods.}
\usage{
subset_members(babase, .adults_only = TRUE)
}
\arguments{
\item{babase}{A DBI connection to the babase database}

\item{.adults_only}{Logical indicating whether to include adults only. Default is TRUE}
}
\value{
A subset of the members table that excludes flagged behavioral observation periods.
}
\description{
Obtain a subset of members table that excludes flagged behavioral observation periods.
}
