% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{dsi}
\alias{dsi}
\title{Calculate DSI variables from individual-year-of-life data}
\usage{
dsi(my_iyol, biograph_l, members_l, focalcount, f_count, grooming_interactions,
  min_cores_days = 60)
}
\arguments{
\item{my_iyol}{Individual-year-of-life data.}

\item{biograph_l}{A local copy of the biograph table}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focalcount}{A subset of focals produced by the function 'subset_focals'}

\item{f_count}{A subset of female counts produced by the function 'subset_females'}

\item{grooming_interactions}{A subset of grooming data produced by the function 'subset_grooming'}

\item{min_cores_days}{The minimum number of coresidence days needed for dyad to be included. Defaults to 60 days.}
}
\value{
The input data with an additional list columsn containing the full DSI subset and the focal DSI variables.
}
\description{
Calculate DSI variables from individual-year-of-life data
}
