% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{get_dyadic_subset}
\alias{get_dyadic_subset}
\title{Obtain dyadic grooming subsets for the animal's year of life}
\usage{
get_dyadic_subset(df, biograph_l, members_l, focalcount, f_count,
  grooming_interactions, min_cores_days = 60)
}
\arguments{
\item{df}{One row individual-year-of-life data}

\item{biograph_l}{A local copy of the biograph table}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focalcount}{A subset of focals produced by the function 'subset_focals'}

\item{f_count}{A subset of female counts produced by the function 'subset_females'}

\item{grooming_interactions}{A subset of grooming data produced by the function 'subset_grooming'}

\item{min_cores_days}{The minimum number of coresidence days needed for dyad to be included. Defaults to 60 days.}
}
\value{
The input row with an additional list column containing the subset
}
\description{
Obtain dyadic grooming subsets for the animal's year of life
}
