% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{get_sci_subset}
\alias{get_sci_subset}
\title{Obtain composite grooming subsets for the animal's year of life}
\usage{
get_sci_subset(df, members_l, focalcount, f_count, grooming_interactions,
  min_res_days = 60)
}
\arguments{
\item{df}{One row individual-year-of-life data}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{focalcount}{A subset of focals produced by the function 'subset_focals'}

\item{f_count}{A subset of female counts produced by the function 'subset_females'}

\item{grooming_interactions}{A subset of grooming data produced by the function 'subset_grooming'}

\item{min_res_days}{The minimum number of residence days needed to be included. Defaults to 60 days.}
}
\value{
The input row with an additional list column containing the subset
}
\description{
Obtain composite grooming subsets for the animal's year of life
}
