% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality-indices.R
\name{subset_members}
\alias{subset_members}
\title{Obtain a subset of members table that excludes behavioral observation gaps.}
\usage{
subset_members(babase)
}
\arguments{
\item{babase}{A DBI connection to the babase database}
}
\value{
A subset of the members table that excludes behavioral observation gaps.
}
\description{
Obtain a subset of members table that excludes behavioral observation gaps.
}
